+++ lib/eve_online/esi/base.rb
@@ -100,6 +100,7 @@ module EveOnline
         @connection ||= Faraday.new { |f|
           f.headers["User-Agent"] = user_agent
           f.headers["If-None-Match"] = _etag if _etag
+          f.headers["Accept"] = "application/json"
           f.authorization :Bearer, token if token
           f.options.read_timeout = _read_timeout
           f.options.open_timeout = _open_timeout