+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## main
+
+* Sync with new ESI
+* Add "Accept" header back
+
 ## 0.38.0
 
 * Replace `net/http` with `faraday`